requirements.txt
# Core API & Server
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20

# Inference Engine
ultralytics==8.4.24
onnxruntime>=1.20.0

# Image Processing (Headless for Server)
opencv-python-headless==4.13.0.92
pillow==12.1.1

# Utilities
pathlib
numpy==2.2.6
PyYAML==6.0.3
requests==2.32.3
supabase==2.11.0
python-dotenv==1.0.1
